+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## _Under development_
 
 - External links now open in new windows.
+- Updated Profile options for Framework v0.5.
+- Updated Issue data to include DOM information.
 - Added support for suspending scans.
 
 ## 0.4.3 _(January 1, 2014)_