+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ChangeLog
 
-## _Under development_
+## 0.5 _(August 29, 2014)_
 
 - External links now open in new windows.
 - Admin can no longer delete self.